link: cleanup link
