+++ static/app/views/settings/organizationApiKeys/organizationApiKeysList.tsx
@@ -2,11 +2,11 @@ import {Fragment} from 'react';
 import {css} from '@emotion/react';
 import styled from '@emotion/styled';
 
+import Confirm from 'sentry/components/confirm';
 import {AlertLink} from 'sentry/components/core/alert/alertLink';
 import {Button} from 'sentry/components/core/button';
 import ExternalLink from 'sentry/components/links/externalLink';
-import Link from 'sentry/components/links/link';
-import LinkWithConfirmation from 'sentry/components/links/linkWithConfirmation';
+import Link, {Anchor} from 'sentry/components/links/link';
 import {PanelTable} from 'sentry/components/panels/panelTable';
 import TextCopyInput from 'sentry/components/textCopyInput';
 import {IconAdd, IconDelete} from 'sentry/icons';
@@ -101,21 +101,25 @@ function OrganizationApiKeysList({
               </TextCopyInput>
 
               <Cell>
-                <LinkWithConfirmation
-                  aria-label={t('Remove API Key')}
-                  className="btn btn-default btn-sm"
+                <Confirm
                   onConfirm={() => onRemove(id)}
                   message={t('Are you sure you want to remove this API key?')}
-                  title={t('Remove API Key?')}
                 >
-                  <IconDelete
-                    size="xs"
-                    css={css`
-                      position: relative;
-                      top: 2px;
-                    `}
-                  />
-                </LinkWithConfirmation>
+                  <Anchor
+                    href="#"
+                    className="btn btn-default btn-sm"
+                    title={t('Remove API Key?')}
+                    aria-label={t('Remove API Key')}
+                  >
+                    <IconDelete
+                      size="xs"
+                      css={css`
+                        position: relative;
+                        top: 2px;
+                      `}
+                    />
+                  </Anchor>
+                </Confirm>
               </Cell>
             </Fragment>
           );
